+++ mountpoint-s3-client/tests/common/mod.rs
@@ -73,15 +73,16 @@ pub fn get_test_client() -> S3CrtClient {
     S3CrtClient::new(S3ClientConfig::new().endpoint_config(endpoint_config)).expect("could not create test client")
 }
 
-pub fn get_test_backpressure_client(initial_read_window: usize) -> S3CrtClient {
+pub fn get_test_backpressure_client(initial_read_window: usize, part_size: Option<usize>) -> S3CrtClient {
     let endpoint_config = EndpointConfig::new(&get_test_region());
-    S3CrtClient::new(
-        S3ClientConfig::new()
-            .endpoint_config(endpoint_config)
-            .read_backpressure(true)
-            .initial_read_window(initial_read_window),
-    )
-    .expect("could not create test client")
+    let mut config = S3ClientConfig::new()
+        .endpoint_config(endpoint_config)
+        .read_backpressure(true)
+        .initial_read_window(initial_read_window);
+    if let Some(part_size) = part_size {
+        config = config.part_size(part_size);
+    }
+    S3CrtClient::new(config).expect("could not create test client")
 }
 
 pub fn get_test_bucket_and_prefix(test_name: &str) -> (String, String) {